+++ assistance_msgs/README.md
@@ -1,3 +1,10 @@
 # Assistance Messages
 
-Defines the interface to the assistance arbitrator and the corresponding assitance interfaces
+Defines the interfaces for:
+
+1. The executive level, the task executor, and the rest of the system to invoke actions or tasks.
+1. Between the executive level and the arbitrator, in the event of an error.
+1. Between the arbitrator and the strategies
+1. The beliefs, to track the semantic state of the system
+
+Additionally, this package contains utilities to enable client libraries to create and parse messages in this package.
