+++ assistance_msgs/package.xml
@@ -54,6 +54,7 @@
   <depend>geometry_msgs</depend>
   <depend>manipulation_actions</depend>
   <depend>rail_manipulation_msgs</depend>
+  <depend>rospy</depend>
   <depend>std_msgs</depend>
 
   <build_depend>message_generation</build_depend>
