+++ predefined_strategy/src/predefined_strategy/recovery_strategies.py
@@ -17,6 +17,7 @@ from assistance_msgs.msg import (RequestAssistanceResult, ExecuteGoal,
 from manipulation_actions.msg import StoreObjectResult, InHandLocalizeResult
 
 from task_executor.actions import get_default_actions
+from assistance_msgs import msg_utils
 
 
 # This class encapsulates the different strategies for recovering from different
@@ -95,7 +96,7 @@ class RecoveryStrategies(object):
         _, beliefs = self._actions.get_beliefs(belief_keys=RecoveryStrategies.BELIEF_KEYS)
 
         # Get the number of times things have failed
-        component_names, num_aborts = RecoveryStrategies.get_number_of_component_aborts(assistance_goal.context)
+        component_names, num_aborts = msg_utils.get_number_of_component_aborts(assistance_goal.context)
 
         # NO COMPETITION: Check for the global recovery abort conditions
         # if len(component_names) > 1 and \
@@ -121,7 +122,7 @@ class RecoveryStrategies(object):
             if assistance_goal.component == 'loop_body_test':
                 rospy.loginfo("Recovery: simply continue")
                 resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-                resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+                resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
             elif assistance_goal.component == 'reposition_recovery_test':
                 rospy.loginfo("Recovery: reposition the base")
@@ -164,14 +165,14 @@ class RecoveryStrategies(object):
             # Alternately, if the base was repositioned, then also retry
             # everything
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
             if (
                 'perceive' in component_names
                 and (assistance_goal.component != 'segment'
                      or num_aborts[component_idx] > 3
                      or RecoveryStrategies.check_contradictory_beliefs(beliefs))
             ):
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'perceive',
                     RequestAssistanceResult.RESUME_RETRY
@@ -181,9 +182,9 @@ class RecoveryStrategies(object):
             rospy.loginfo("Recovery: wait before redetect")
             self._actions.wait(duration=0.5)
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
             if 'pick_kit_task' in component_names and num_aborts[-1] >= 3:
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'pick_kit_task',
                     RequestAssistanceResult.RESUME_RETRY
@@ -193,9 +194,9 @@ class RecoveryStrategies(object):
             rospy.loginfo("Recovery: wait and try to redetect")
             self._actions.wait(duration=0.5)
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
             if 'detect_schunk_pose_task' in component_names:
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'detect_schunk_pose_task',
                     RequestAssistanceResult.RESUME_RETRY
@@ -245,12 +246,12 @@ class RecoveryStrategies(object):
                 )
 
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
             # If this is a pick, we want to retry segmentation
             if assistance_goal.component == 'pick' and 'perceive_pick' in component_names:
                 rospy.loginfo("Recovery: restarting pick with perception")
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'perceive_pick',
                     RequestAssistanceResult.RESUME_RETRY
@@ -258,7 +259,7 @@ class RecoveryStrategies(object):
 
         elif assistance_goal.component == 'move':
             rospy.loginfo("Recovery: reposition, then retry move to goal pose")
-            component_context = RecoveryStrategies.get_final_component_context(assistance_goal.context)
+            component_context = msg_utils.get_final_component_context(assistance_goal.context)
             self._actions.move_planar(angular_amount=np.pi / 10)
             self._actions.wait(duration=0.5)
             self._actions.move_planar(angular_amount=-1 * np.pi / 10)
@@ -278,7 +279,7 @@ class RecoveryStrategies(object):
                 )
 
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif assistance_goal.component == 'reposition':
             rospy.loginfo("Recovery: wiggle back and forth, then retry reposition")
@@ -290,15 +291,15 @@ class RecoveryStrategies(object):
             self._actions.wait(duration=0.5)
 
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif assistance_goal.component == 'move_backward':
             rospy.loginfo("Recovery: move forward a little bit")
-            component_context = RecoveryStrategies.get_final_component_context(assistance_goal.context)
+            component_context = msg_utils.get_final_component_context(assistance_goal.context)
             goal_amount = component_context.get('goal', 0.0)
             self._actions.move_backward(amount=-goal_amount/2)
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif (
             assistance_goal.component == 'store_object'
@@ -308,9 +309,9 @@ class RecoveryStrategies(object):
             self._actions.wait(duration=0.5)
             self._actions.load_static_octomap()
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
-            component_context = RecoveryStrategies.get_final_component_context(assistance_goal.context)
+            component_context = msg_utils.get_final_component_context(assistance_goal.context)
 
             # If this is store object and the result indicates that it exited
             # with a verify grasp failure, then we should redo that object's
@@ -322,7 +323,7 @@ class RecoveryStrategies(object):
                 and component_context['result'].error_code == StoreObjectResult.ABORTED_ON_GRASP_VERIFICATION
             ):
                 rospy.loginfo("Recovery: retrying pick-and-place")
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'pick_place_in_kit',
                     RequestAssistanceResult.RESUME_RETRY
@@ -338,7 +339,7 @@ class RecoveryStrategies(object):
                 and component_context['result'].error_code == StoreObjectResult.ABORTED_ON_GRASP_VERIFICATION
             ):
                 rospy.loginfo("Recovery: have to retry filling the kit")
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'fill_kit',
                     RequestAssistanceResult.RESUME_RETRY
@@ -354,7 +355,7 @@ class RecoveryStrategies(object):
                 and component_context['result'].error_code == InHandLocalizeResult.ABORTED_ON_POSE_CHECK
             ):
                 rospy.loginfo("Recovery: dropping off the large gear before retrying pick-and-place")
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'pick_insert_gear_in_schunk',
                     RequestAssistanceResult.RESUME_RETRY
@@ -365,27 +366,27 @@ class RecoveryStrategies(object):
             rospy.loginfo("Recovery: move arm to verify, then retry the pick")
             self._actions.load_static_octomap()
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif assistance_goal.component == 'place_kit_base':
             rospy.loginfo("Recovery: retry the place")
             self._actions.load_static_octomap()
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif assistance_goal.component == 'verify_grasp':
             rospy.loginfo("Recovery: object dropped, retry the pick")
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
             if 'pick_place_in_kit' in component_names:
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'pick_place_in_kit',
                     RequestAssistanceResult.RESUME_RETRY
                 )
             elif 'pick_place_kit_on_robot' in component_names:
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'pick_place_kit_on_robot',
                     RequestAssistanceResult.RESUME_RETRY
@@ -393,7 +394,7 @@ class RecoveryStrategies(object):
             elif 'pick_insert_gear_in_schunk' in component_names:
                 rospy.loginfo("Recovery: pull back and then try again")
                 self._actions.move_backward(amount=0.4)
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'pick_insert_gear_in_schunk',
                     RequestAssistanceResult.RESUME_RETRY
@@ -402,7 +403,7 @@ class RecoveryStrategies(object):
                 rospy.loginfo("Recovery: open gripper and move backwards")
                 self._actions.gripper(command="open")
                 self._actions.move_backward(amount=0.4)
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'pick_from_schunk_task',
                     RequestAssistanceResult.RESUME_RETRY
@@ -410,7 +411,7 @@ class RecoveryStrategies(object):
             elif 'remove_place_gear_in_kit' in component_names:
                 rospy.loginfo("Recovery: large gear dropped somewhere. Restarting the whole task")
                 if 'fill_kit' in component_names:
-                    resume_context = RecoveryStrategies.set_task_hint_in_context(
+                    resume_context = msg_utils.set_task_hint_in_context(
                         resume_context,
                         'fill_kit',
                         RequestAssistanceResult.RESUME_RETRY
@@ -425,9 +426,9 @@ class RecoveryStrategies(object):
             rospy.loginfo("Recovery: could not plan to approach pose, clearing octomap and retrying")
             self._actions.load_static_octomap()
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
             if 'arm_approach_schunk_task' in component_names:
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'arm_approach_schunk_task',
                     RequestAssistanceResult.RESUME_RETRY
@@ -446,7 +447,7 @@ class RecoveryStrategies(object):
                     name="reposition_recovery_task",
                     params=pickle.dumps(goal_params)
                 )
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'insert_in_schunk_task',
                     RequestAssistanceResult.RESUME_PREVIOUS
@@ -456,21 +457,21 @@ class RecoveryStrategies(object):
             rospy.loginfo("Recovery: wait and simply retry")
             self._actions.wait(duration=0.5)
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif assistance_goal.component == 'schunk':
             rospy.loginfo("Recovery: wait and then try the schunk again")
             self._actions.wait(duration=5.0)
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif assistance_goal.component == 'schunk_insertion':
             rospy.loginfo("Recovery: move backwards and then try the task again")
             self._actions.move_backward(amount=0.4)
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
             if 'insert_in_schunk_task' in component_names:
-                resume_context = RecoveryStrategies.set_task_hint_in_context(
+                resume_context = msg_utils.set_task_hint_in_context(
                     resume_context,
                     'insert_in_schunk_task',
                     RequestAssistanceResult.RESUME_RETRY
@@ -479,7 +480,7 @@ class RecoveryStrategies(object):
         elif assistance_goal.component == 'schunk_gripper_pullback':
             rospy.loginfo("Recovery: retry pullback")
             resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-            resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+            resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
 
         elif assistance_goal.component == 'grasp_schunk_gear':
             _, action_result = self._actions.verify_grasp()
@@ -487,9 +488,9 @@ class RecoveryStrategies(object):
                 rospy.loginfo("Recovery: No gear in hand. Retry")
                 self._actions.move_backward(amount=0.4)
                 resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-                resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+                resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
                 if 'pick_from_schunk_task' in component_names:
-                    resume_context = RecoveryStrategies.set_task_hint_in_context(
+                    resume_context = msg_utils.set_task_hint_in_context(
                         resume_context,
                         'pick_from_schunk_task',
                         RequestAssistanceResult.RESUME_RETRY
@@ -497,9 +498,9 @@ class RecoveryStrategies(object):
             else:
                 rospy.loginfo("Recovery: gear in hand. Move to the next step")
                 resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-                resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+                resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
                 if 'pick_from_schunk_task' in component_names:
-                    resume_context = RecoveryStrategies.set_task_hint_in_context(
+                    resume_context = msg_utils.set_task_hint_in_context(
                         resume_context,
                         'pick_from_schunk_task',
                         RequestAssistanceResult.RESUME_NEXT
@@ -514,8 +515,8 @@ class RecoveryStrategies(object):
                 # We had the gear and at some point we lost it, have to restart
                 if 'fill_kit' in component_names:
                     resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-                    resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
-                    resume_context = RecoveryStrategies.set_task_hint_in_context(
+                    resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
+                    resume_context = msg_utils.set_task_hint_in_context(
                         resume_context,
                         'fill_kit',
                         RequestAssistanceResult.RESUME_RETRY
@@ -528,9 +529,9 @@ class RecoveryStrategies(object):
             else:
                 rospy.loginfo("Recovery: gear in hand. Try to move away")
                 resume_hint = RequestAssistanceResult.RESUME_CONTINUE
-                resume_context = RecoveryStrategies.create_continue_result_context(assistance_goal.context)
+                resume_context = msg_utils.create_continue_result_context(assistance_goal.context)
                 if 'remove_place_gear_in_kit' in component_names:
-                    resume_context = RecoveryStrategies.set_task_hint_in_context(
+                    resume_context = msg_utils.set_task_hint_in_context(
                         resume_context,
                         'remove_place_gear_in_kit',
                         RequestAssistanceResult.RESUME_NEXT
@@ -543,104 +544,6 @@ class RecoveryStrategies(object):
             resume_context
         ))
         return execute_goal, resume_hint, resume_context
-
-    @staticmethod
-    def get_final_component_context(goal_context):
-        """
-        Get the context of the last component in the context chain
-        """
-        if len(goal_context.get('context', {})) > 0:
-            return RecoveryStrategies.get_final_component_context(goal_context['context'])
-        else:
-            return goal_context
-
-    @staticmethod
-    def get_number_of_component_aborts(goal_context):
-        """
-        Given the hierarchy of tasks in the goal context, obtain a vector of the
-        number of failures in each part of the component of the hierarchy. The
-        first index maps to the highest level of the hierarcy and the last
-        index maps to the lowest level of the hierarchy.
-
-        Args:
-            goal_context (dict) : the goal context
-
-        Returns:
-            (tuple):
-                - component_names (list) a list of component names from highest \
-                    in the task hierarchy to the lowest
-                - num_aborts (list) a list of the number of times each \
-                    component in component_names aborted
-        """
-        component_names = []
-        num_aborts = []
-        sub_context = goal_context
-
-        while sub_context is not None and isinstance(sub_context, dict):
-            component_names.append(sub_context.get('task') or sub_context.get('action'))
-            num_aborts.append(sub_context.get('num_aborts'))
-            sub_context = sub_context.get('context')
-
-        # Return the lists
-        return (component_names, num_aborts,)
-
-    @staticmethod
-    def create_continue_result_context(goal_context):
-        """
-        Given the context of a ``assistance_msgs/RequestAssistanceGoal``
-        return a dictionary for a ``assistance_msgs/RequestAssistanceResult``
-        context that indicates :const:`RequestAssistanceResult.RESUME_CONTINUE`
-
-        Args:
-            goal_context (dict) : the goal context
-
-        Return:
-            (dict) : the result context
-        """
-        if 'task' in goal_context:
-            return {
-                'task': goal_context['task'],
-                'step_idx': goal_context['step_idx'],
-                'resume_hint': RequestAssistanceResult.RESUME_CONTINUE,
-                'context': RecoveryStrategies.create_continue_result_context(goal_context['context']),
-            }
-        else:
-            return {}
-
-    @staticmethod
-    def set_task_hint_in_context(result_context, task_name, resume_hint):
-        """
-        Given a result context dictionary, mark the desired task name with the
-        desired resume hint
-
-        Args:
-            result_context (dict) : the result context, possibly created by
-                :meth:`create_continue_result_context`
-            task_name (str) : the name of the task
-            resume_hint (uint8) : A ``assistance_msgs/RequestAssistanceResult`` \
-                resume_hint constant for the task's resume hint
-
-        Returns:
-            (dict) : a result context dictionary with the task set to the \
-                desired resume_hint. Note: we do not copy, so the incoming arg \
-                might also get affected
-
-        Raises:
-            KeyError : if :data:`task_name` is not found in the context
-        """
-
-        # Error checking
-        if 'task' not in result_context:
-            raise KeyError("Expected a result context for tasks. Not found in {}!".format(result_context))
-
-        # If this is not the task we want, then continue on to its context.
-        # Otherwise, mark this task as the one we want to update and return
-        if result_context['task'] == task_name:
-            result_context['resume_hint'] = resume_hint
-        else:
-            result_context['context'] = RecoveryStrategies.set_task_hint_in_context(result_context['context'], task_name, resume_hint)
-
-        return result_context
 
     @staticmethod
     def check_contradictory_beliefs(beliefs):
