+++ task_executor/src/task_executor/server.py
@@ -8,6 +8,7 @@ import pickle
 import rospy
 import actionlib
 
+from assistance_msgs import msg_utils
 from sound_interface import SoundClient
 from task_executor.abstract_step import AbstractStep
 from task_executor.actions import get_default_actions
@@ -150,7 +151,7 @@ class TaskServer(object):
                 # If the task has been preempted, then stop executing it
                 if task.is_preempted():
                     rospy.logwarn("Task {}: PREEMPTED. Context: {}".format(
-                        task.name, Task.pprint_variables(variables)
+                        task.name, msg_utils.pprint_context(variables)
                     ))
                     result.variables = pickle.dumps(variables)
                     self._server.set_preempted(result)
@@ -159,7 +160,7 @@ class TaskServer(object):
                 # If the task has failed, request assistance and resume based
                 if task.is_aborted():
                     rospy.logerr("Task {}: FAIL. Context: {}".format(
-                        task.name, Task.pprint_variables(variables)
+                        task.name, msg_utils.pprint_context(variables)
                     ))
                     request_assistance = True
 
@@ -255,7 +256,7 @@ class TaskServer(object):
         # Check to see if the task aborted
         if task.is_aborted():
             rospy.logerr("Task {}: FAIL. Context: {}".format(
-                task.name, Task.pprint_variables(variables)
+                task.name, msg_utils.pprint_context(variables)
             ))
             result.variables = pickle.dumps(variables)
             self._server.set_aborted(result)
